requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apt-clone==0.2.1
apturl==0.5.2
bcrypt==4.1.2
beautifulsoup4==4.10.0
billiard==4.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
celery==5.3.6
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cssselect==1.2.0
cssutils==2.9.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
ecdsa==0.18.0
emails==0.6
et-xmlfile==1.1.0
exceptiongroup==1.2.0
eyeD3==0.8.10
fastapi==0.109.0
filelock==3.6.0
greenlet==3.0.3
grpcio==1.30.2
h11==0.14.0
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kombu==5.3.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
nemo-emblems==5.8.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.6.3
numpy==1.26.1
oauthlib==3.2.0
onboard==1.4.1
openpyxl==3.1.2
packaging==21.3
PAM==0.4.2
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
premailer==3.10.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pycairo==1.20.1
pycel==1.0b30
pycups==2.0.1
pycurl==7.44.1
pydantic==2.5.3
pydantic_core==2.14.6
pyelftools==0.27
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.6
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
pyxlsx==1.1.3
PyYAML==5.4.1
raven==6.10.0
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rsa==4.9
ruamel.yaml==0.18.2
ruamel.yaml.clib==0.2.8
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.35.1
systemd-python==234
tenacity==8.2.3
tinycss2==1.1.1
tldextract==3.1.2
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
uvicorn==0.27.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
